+++ src/receiver.rs
@@ -1,4 +1,3 @@
-use crate::TaskError;
 use futures::task::AtomicWaker;
 use std::sync::{
     atomic::{AtomicBool, Ordering},
@@ -11,12 +10,12 @@ use tokio::sync::oneshot::{self};
 pub struct AsyncReceiver<T> {
     pub(crate) received: Arc<AtomicBool>,
     pub(crate) waker: Arc<AtomicWaker>, // Waker to wake the sender
-    pub(crate) receiver: oneshot::Receiver<Result<T, TaskError>>,
+    pub(crate) receiver: oneshot::Receiver<T>,
 }
 
 impl<T> AsyncReceiver<T> {
     /// Poll the current thread waiting for the async result.
-    pub fn try_recv(&mut self) -> Option<Result<T, TaskError>> {
+    pub fn try_recv(&mut self) -> Option<T> {
         match self.receiver.try_recv() {
             Ok(t) => {
                 self.receiver.close();