+++ src/task.rs
@@ -16,12 +16,31 @@ use std::{
 };
 use tokio::sync::oneshot;
 
+/// A wrapper type around an async future with a timeout. The future may be executed
+/// asynchronously by an [`TimedTaskRunner`](crate::TimedTaskRunner) or
+/// [`TimedTaskPool`](crate::TimedTaskPool) bevy system parameter.
+pub struct TimedAsyncTask<T: ConditionalSend> {
+    fut: Pin<Box<dyn ConditionalSendFuture<Output = T> + 'static>>,
+    timeout: Duration,
+}
+
+impl<T> Debug for TimedAsyncTask<T>
+where
+    T: Debug + Send,
+{
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("TimedAsyncTask")
+            .field("fut", &"<future>")
+            .field("timeout", &self.timeout)
+            .finish()
+    }
+}
+
 /// A wrapper type around an async future. The future may be executed
-/// asynchronously by an [`AsyncTaskRunner`](crate::AsyncTaskRunner) or
-/// [`AsyncTaskPool`](crate::AsyncTaskPool) bevy system parameter.
+/// asynchronously by an [`TaskRunner`](crate::TaskRunner) or
+/// [`TaskPool`](crate::TaskPool) bevy system parameter.
 pub struct AsyncTask<T: ConditionalSend> {
     fut: Pin<Box<dyn ConditionalSendFuture<Output = T> + 'static>>,
-    timeout: Option<Duration>,
 }
 
 impl<T> Debug for AsyncTask<T>
@@ -31,7 +50,6 @@ where
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("AsyncTask")
             .field("fut", &"<future>")
-            .field("timeout", &self.timeout)
             .finish()
     }
 }
@@ -40,20 +58,24 @@ impl<T> AsyncTask<T>
 where
     T: ConditionalSend + 'static,
 {
-    /// Never resolves to a value or finishes.
+    /// Create an async task from a future.
+    pub fn new<F>(fut: F) -> Self
+    where
+        F: ConditionalSendFuture<Output = T> + 'static,
+        F::Output: ConditionalSend + 'static,
+    {
+        Self { fut: Box::pin(fut) }
+    }
+
+    /// Never resolves to a value or becomes ready.
     pub fn pending() -> Self {
         Self::new(pending())
     }
-}
 
-impl<T> AsyncTask<T>
-where
-    T: ConditionalSend,
-{
-    /// Build the task into a runnable future and receiver.
+    /// Split the task into a runnable future and receiver.
     /// This is a low-level operation and only useful for specific needs.
     #[must_use]
-    pub fn build(self) -> (Pin<Box<impl Future<Output = ()>>>, AsyncReceiver<T>) {
+    pub fn split(self) -> (Pin<Box<impl Future<Output = ()>>>, AsyncReceiver<T>) {
         let (tx, rx) = oneshot::channel();
         let waker = Arc::new(AtomicWaker::new());
         let received = Arc::new(AtomicBool::new(false));
@@ -62,19 +84,9 @@ where
             let received = received.clone();
             async move {
                 #[cfg(target_arch = "wasm32")]
-                let result = if let Some(dur) = self.timeout {
-                    timeout(dur, self.fut).await.map_err(TaskError::Timeout)
-                } else {
-                    Ok(self.fut.await)
-                };
+                let result = self.fut.await;
                 #[cfg(not(target_arch = "wasm32"))]
-                let result = if let Some(dur) = self.timeout {
-                    timeout(dur, self.fut.compat())
-                        .await
-                        .map_err(TaskError::Timeout)
-                } else {
-                    Ok(self.fut.compat().await)
-                };
+                let result = self.fut.compat().await;
 
                 if let Ok(()) = tx.send(result) {
                     // Wait for the receiver to get the result before dropping.
@@ -99,36 +111,12 @@ where
         (fut, receiver)
     }
 
-    /// Create an async task from a future.
-    pub fn new<F>(fut: F) -> Self
-    where
-        F: ConditionalSendFuture<Output = T> + 'static,
-        F::Output: ConditionalSend + 'static,
-    {
-        Self {
-            fut: Box::pin(fut),
-            timeout: None,
-        }
-    }
-
-    /// Create an async task from a future with a timeout.
-    pub fn new_with_timeout<F>(dur: Duration, fut: F) -> Self
-    where
-        F: ConditionalSendFuture<Output = T> + 'static,
-        F::Output: ConditionalSend + 'static,
-    {
-        Self {
-            fut: Box::pin(fut),
-            timeout: Some(dur),
-        }
-    }
-
-    /// Replace the timeout for this task.
+    /// Add a timeout for this task.
     #[must_use]
-    pub fn with_timeout(self, dur: Duration) -> Self {
-        Self {
-            timeout: Some(dur),
-            ..self
+    pub fn with_timeout(self, dur: Duration) -> TimedAsyncTask<T> {
+        TimedAsyncTask {
+            fut: self.fut,
+            timeout: dur,
         }
     }
 }
@@ -140,6 +128,99 @@ where
 {
     fn from(value: Fnc) -> Self {
         Self::new(value)
+    }
+}
+
+impl<T> TimedAsyncTask<T>
+where
+    T: ConditionalSend + 'static,
+{
+    /// Create an async task from a future with a timeout.
+    pub fn new<F>(dur: Duration, fut: F) -> Self
+    where
+        F: ConditionalSendFuture<Output = T> + 'static,
+        F::Output: ConditionalSend + 'static,
+    {
+        Self {
+            fut: Box::pin(fut),
+            timeout: dur,
+        }
+    }
+
+    /// Never resolves to a value or becomes ready.
+    pub fn pending() -> Self {
+        Self::new(crate::DEFAULT_TIMEOUT, pending())
+    }
+
+    /// Split the task into a runnable future and receiver.
+    /// This is a low-level operation and only useful for specific needs.
+    #[must_use]
+    pub fn split(
+        self,
+    ) -> (
+        Pin<Box<impl Future<Output = ()>>>,
+        AsyncReceiver<Result<T, TaskError>>,
+    ) {
+        let (tx, rx) = oneshot::channel();
+        let waker = Arc::new(AtomicWaker::new());
+        let received = Arc::new(AtomicBool::new(false));
+        let fut = {
+            let waker = waker.clone();
+            let received = received.clone();
+            async move {
+                #[cfg(target_arch = "wasm32")]
+                let result = timeout(self.timeout, self.fut)
+                    .await
+                    .map_err(TaskError::Timeout);
+                #[cfg(not(target_arch = "wasm32"))]
+                let result = timeout(self.timeout, self.fut.compat())
+                    .await
+                    .map_err(TaskError::Timeout);
+
+                if let Ok(()) = tx.send(result) {
+                    // Wait for the receiver to get the result before dropping.
+                    futures::future::poll_fn(|cx| {
+                        waker.register(cx.waker());
+                        if received.load(Ordering::Relaxed) {
+                            Poll::Ready(())
+                        } else {
+                            Poll::Pending::<()>
+                        }
+                    })
+                    .await;
+                }
+            }
+        };
+        let fut = Box::pin(fut);
+        let receiver = AsyncReceiver {
+            received,
+            waker,
+            receiver: rx,
+        };
+        (fut, receiver)
+    }
+
+    /// Replace the timeout for this task.
+    #[must_use]
+    pub fn with_timeout(mut self, dur: Duration) -> Self {
+        self.timeout = dur;
+        self
+    }
+
+    /// Remove the timeout for this task.
+    #[must_use]
+    pub fn without_timeout(self) -> AsyncTask<T> {
+        AsyncTask { fut: self.fut }
+    }
+}
+
+impl<T, Fnc> From<Fnc> for TimedAsyncTask<T>
+where
+    Fnc: ConditionalSendFuture<Output = T> + 'static,
+    Fnc::Output: ConditionalSend + 'static,
+{
+    fn from(value: Fnc) -> Self {
+        Self::new(crate::DEFAULT_TIMEOUT, value)
     }
 }
 
@@ -171,7 +252,7 @@ mod test {
     #[tokio::test]
     async fn test_try_recv() {
         let task = AsyncTask::new(async move { 5 });
-        let (fut, mut rx) = task.build();
+        let (fut, mut rx) = task.split();
 
         assert_eq!(None, rx.try_recv());
 
@@ -186,7 +267,7 @@ mod test {
             select! {
                 _ = (&mut fetch).fuse() => {
                     if let Some(v) = rx.try_recv() {
-                        assert_eq!(5, v.unwrap());
+                        assert_eq!(5, v);
                         break 'result;
                     } else {
                         // Reset the clock
@@ -200,8 +281,8 @@ mod test {
 
     #[tokio::test]
     async fn test_timeout() {
-        let task = AsyncTask::new_with_timeout(Duration::from_millis(5), pending::<()>());
-        let (fut, mut rx) = task.build();
+        let task = TimedAsyncTask::new(Duration::from_millis(5), pending::<()>());
+        let (fut, mut rx) = task.split();
 
         assert_eq!(None, rx.try_recv());
 
@@ -234,8 +315,8 @@ mod test {
 
     #[tokio::test]
     async fn test_with_timeout() {
-        let task = AsyncTask::new_with_timeout(Duration::from_millis(5), pending::<()>());
-        let (fut, mut rx) = task.build();
+        let task = TimedAsyncTask::new(Duration::from_millis(5), pending::<()>());
+        let (fut, mut rx) = task.split();
 
         assert_eq!(None, rx.try_recv());
 
@@ -301,7 +382,7 @@ mod test {
     #[wasm_bindgen_test]
     async fn test_try_recv() {
         let task = AsyncTask::new(async move { 5 });
-        let (fut, mut rx) = task.build();
+        let (fut, mut rx) = task.split();
 
         assert_eq!(None, rx.try_recv());
 
@@ -316,13 +397,13 @@ mod test {
         });
 
         // Spawn
-        assert_eq!(Some(Ok(5)), rx.try_recv());
+        assert_eq!(Some(5), rx.try_recv());
     }
 
     #[wasm_bindgen_test]
     async fn test_timeout() {
-        let task = AsyncTask::new_with_timeout(Duration::from_millis(5), pending::<()>());
-        let (fut, mut rx) = task.build();
+        let task = TimedAsyncTask::<()>::pending().with_timeout(Duration::from_millis(5));
+        let (fut, mut rx) = task.split();
 
         assert_eq!(None, rx.try_recv());
 
@@ -345,8 +426,8 @@ mod test {
 
     #[wasm_bindgen_test]
     async fn test_with_timeout() {
-        let task = AsyncTask::new_with_timeout(Duration::from_millis(5), pending::<()>());
-        let (fut, mut rx) = task.build();
+        let task = TimedAsyncTask::<()>::pending().with_timeout(Duration::from_millis(5));
+        let (fut, mut rx) = task.split();
 
         assert_eq!(None, rx.try_recv());